# Refactor code structure, change UUID class and file to UUIDs (avoid same naming accross libraries), and enhance JSON serialization in Sensors_Processes. Update configuration settings and library dependencies.  # All base functionality implemented, ready for testing and further development.  # Data sent from Azure IOT Hub isn't being processed yet, but the structure is in place for future enhancements.  # Small callibration may be needed for MPU6050, MAX30102, DS18B20 sensors.  # For testing purpose, the uploading routine is set to every 2nd sensor reading ittreation + sleep.

+++ lib/MAX30102/MAX30102.cpp
@@ -154,7 +154,7 @@ bool MAX30102::readRegister(uint8_t reg_addr, uint8_t *value) {
         xprintln("I2C: Failed to send reg addr for read");
         return false;
     }
-    if (Wire.requestFrom(MAX30102_ADDRESS, (uint8_t)1) == 1) {
+    if (Wire.requestFrom((uint8_t)MAX30102_ADDRESS, (uint8_t)1) == 1) {
         *value = Wire.read();
         return true;
     }
@@ -170,7 +170,7 @@ bool MAX30102::readBurst(uint8_t start_reg_addr, uint8_t *buffer, uint8_t count)
         xprintln("I2C: Failed to send reg addr for burst read");
         return false;
     }
-    if (Wire.requestFrom(MAX30102_ADDRESS, count) == count) {
+    if (Wire.requestFrom((uint8_t)MAX30102_ADDRESS, count) == count) {
         for (uint8_t i = 0; i < count; i++) {
             buffer[i] = Wire.read();
         }
